Update Switch.cs
Added functionality to "traps." Basically, this script will let us attach to one specific "object" that is always a "switch" behavior. When we want objects to disappear/reappear (or later on change it to open/close doors) then we call the "TriggerSwitch()" and if we want to disable "Traps" that deal Damage over time or death traps (needs to be implemented), we can call "Activate()" that will go send a message to the trap script ("DamageBurn()") and activate it's "Toggle()" which makes it enabled or disabled.