Spotlight Class created
This will make it so we have only one spotlight running at once during
gameplay. I've also created a player prefab and will upload it, but I
don't know if it will appear as it should. I can't get the models to
appear under main camera, where all the objects are:

Player > MainCamera > Spotlight : MoonGem : SunGem : Canteen :
Flashlight

is how it should be. Spotlight is not under any item, it stands alone
and has this script attached, then in the inspector just attach the
Spotlight that you want to reference to the Spotlight variable in the
inspector and you're done. :D