Create MoveObject.cs
This script will let us use switches to move platforms/objects/doors. For now, it only let's us move in one direction (forward/backward/up/down/left/right). When a switch is triggered, the object will move in X/Y/Z direction, based upon the MaxDistance given via inspector. When it reaches that destination, it will move. I haven't made it so that the switch won't trigger the object to move again, but that just is a simple tweak via Switch.cs which I will try to make. 

The purpose of this script is to let us use it for sliding doors, and other objects we may find a use for it. Like in the first level, when the "door" unlocks, I figured it could be a secret passage behind the Church, which gets uncovered when the player opens up the door.